View for Students
Created the view for the viewing students functionality for both lecturer and Guidance. Only guidance will have the option to add a new student. Includes a search bar to quickly find students and options to delete, edit or see their full details although this has been commented out for now.